requirements.txt
Flask==0.10.1
Flask-Bootstrap==3.3.5.7
Jinja2==2.8
MarkupSafe==0.23
Werkzeug==0.11.9
click==6.6
dominate==2.2.0
gunicorn==19.4.5
itsdangerous==0.24
ordereddict==1.1
python-dotenv==0.4.0
requests==2.10.0
visitor==0.1.2
wsgiref==0.1.2
